requirement.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
appdirs==1.4.4
astor==0.8.1
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
driver==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
mysql-connector-python==8.0.33
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
protobuf==3.20.3
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==7.3.1
pytest-ast-transformer==1.0.3
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-soft-assertions==0.1.2
pytest-xdist==3.3.1
requests==2.30.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
texttable==1.6.7
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==2.0.2
webdrivermanager==0.10.0
wsproto==1.2.0
